requirements.txt
python-telegram-bot==21.0
openai==1.40.0
langchain==0.2.16
langchain-openai==0.1.23
httpx==0.27.0
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0